changed status_manager to state_manager; resolved type directory
