--- vda5050_core/include/vda5050_core/state/info_level.hpp
@@ -1,68 +1,0 @@
-/*
- * Copyright (C) 2025 ROS-Industrial Consortium Asia Pacific
- * Advanced Remanufacturing and Technology Centre
- * A*STAR Research Entities (Co. Registration No. 199702110H)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef VDA5050_CORE__STATE__INFO_LEVEL_HPP_
-#define VDA5050_CORE__STATE__INFO_LEVEL_HPP_
-
-#include <nlohmann/json.hpp>
-
-namespace vda5050_core {
-
-namespace msg {
-
-/// @enum InfoLevel
-/// @brief Information for debugging or visualization
-enum class InfoLevel
-{
-  /// @brief used for debugging
-  DEBUG,
-
-  /// @brief used for visualization
-  INFO
-};
-
-using json = nlohmann::json;
-
-inline void to_json(json& j, const InfoLevel& level)
-{
-  switch (level)
-  {
-    case InfoLevel::DEBUG:
-      j = "DEBUG";
-      break;
-    case InfoLevel::INFO:
-      j = "INFO";
-      break;
-  }
-}
-
-inline void from_json(const json& j, InfoLevel& level)
-{
-  const std::string& s = j.get<std::string>();
-  if (s == "DEBUG")
-    level = InfoLevel::DEBUG;
-  else if (s == "INFO")
-    level = InfoLevel::INFO;
-  else
-    throw std::invalid_argument("Invalid InfoLevel value: " + s);
-}
-
-}  // namespace msg
-}  // namespace vda5050_core
-
-#endif  // VDA5050_CORE__STATE__INFO_LEVEL_HPP_
